Add interfaces GameState and SelectionState, implement them, add SelectionManager
The main ideas behind states are: first, to copy data if and only if it is changed; second, to provide readonly access to copies; third, to guarantee that copies are not modified while you're working with them. It's supposed to be good for rendering and state-based logic.

Too many changes in one commit, my bad.