Add FieldStructure
Basically it's a non-static storage of all the cellCodes. The main idea is to have field structure separated from field data so that there's a defined way to say "give me link X" and there're easy opportunities to make it into "link M might be in one of ten states, give me the right one".